refactor: extracted data classes and enums from GameBoard Screen
