+++ app/src/main/java/com/codenames/frontend/data/model/GameCard.kt
@@ -1,0 +1,9 @@
+package com.codenames.frontend.data.model
+
+import com.codenames.frontend.data.model.enums.CardType
+
+data class GameCard(
+    val word: String,
+    val type: CardType,
+    val revealed: Boolean = false,
+)
