+++ app/src/main/java/com/codenames/frontend/data/model/enums/CardType.kt
@@ -1,0 +1,8 @@
+package com.codenames.frontend.data.model.enums
+
+enum class CardType {
+    BLUE,
+    RED,
+    NEUTRAL,
+    ASSASSIN,
+}
